+++ table/src/test/java/tech/ydb/table/impl/pool/FutureHelper.java
@@ -73,6 +73,4 @@ public class FutureHelper {
             Assert.assertNotNull("Test interrupted", ex);
         }
     }
-
-
 }